Add signature to allow user to specify overwrite
This change adds additional signatures to give the user the ability to
allow overwrite of a directory if needed during the extraction process
Test cases are included that demonstrate the existing exception
and test usage of the overwrite flag.

This version moves the boolean parameter to a separate function
signature and removes the default parameter.

Fix #8146